Rename file names back to introduce less diff
