+++ rag_experiment_accelerator/doc_loader/jsonLoader.py
@@ -1,9 +1,9 @@
 from typing import Iterable
 
-from rag_experiment_accelerator.doc_loader.custom_json_loader import (
+from rag_experiment_accelerator.doc_loader.customJsonLoader import (
     CustomJSONLoader,
 )
-from rag_experiment_accelerator.doc_loader.structured_loader import (
+from rag_experiment_accelerator.doc_loader.structuredLoader import (
     load_structured_files,
 )
 from rag_experiment_accelerator.utils.logging import get_logger
