+++ rag_experiment_accelerator/run/index.py
@@ -1,7 +1,7 @@
 from rag_experiment_accelerator.config import Config
 from rag_experiment_accelerator.config.index_config import IndexConfig
 from rag_experiment_accelerator.config.environment import Environment
-from rag_experiment_accelerator.doc_loader.document_loader import load_documents
+from rag_experiment_accelerator.doc_loader.documentLoader import load_documents
 from rag_experiment_accelerator.ingest_data.acs_ingest import upload_data
 from rag_experiment_accelerator.init_Index.create_index import create_acs_index
 from rag_experiment_accelerator.nlp.preprocess import Preprocess
